Reorganised interfaces, putting them where they are used.
