+++ src/main/io/displayport_msp.c
@@ -34,11 +34,10 @@
 
 #include "drivers/display.h"
 
-#include "interface/msp.h"
-#include "interface/msp_protocol.h"
-
 #include "io/displayport_msp.h"
 
+#include "msp/msp.h"
+#include "msp/msp_protocol.h"
 #include "msp/msp_serial.h"
 
 // no template required since defaults are zero
