+++ src/main/msp/msp_serial.h
@@ -21,7 +21,7 @@
 #pragma once
 
 #include "drivers/time.h"
-#include "interface/msp.h"
+#include "msp/msp.h"
 
 // Each MSP port requires state and a receive buffer, revisit this default if someone needs more than 3 MSP ports.
 #define MAX_MSP_PORT_COUNT 3
