+++ src/main/sensors/current.c
@@ -254,7 +254,8 @@ void currentMeterESCReadMotor(uint8_t motorNumber, currentMeter_t *meter)
 
 #ifdef USE_MSP_CURRENT_METER
 #include "common/streambuf.h"
-#include "interface/msp_protocol.h"
+
+#include "msp/msp_protocol.h"
 #include "msp/msp_serial.h"
 
 currentMeterMSPState_t currentMeterMSPState;
