+++ src/main/telemetry/crsf.h
@@ -24,7 +24,8 @@
 #include <stdint.h>
 
 #include "common/time.h"
-#include "interface/crsf_protocol.h"
+
+#include "rx/crsf_protocol.h"
 
 #define CRSF_MSP_RX_BUF_SIZE 128
 #define CRSF_MSP_TX_BUF_SIZE 128
