+++ src/test/unit/cli_unittest.cc
@@ -27,10 +27,9 @@ extern "C" {
     #include "platform.h"
     #include "target.h"
     #include "build/version.h"
+    #include "cli/cli.h"
+    #include "cli/settings.h"
     #include "config/feature.h"
-    #include "pg/pg.h"
-    #include "pg/pg_ids.h"
-    #include "pg/rx.h"
     #include "drivers/buf_writer.h"
     #include "drivers/vtx_common.h"
     #include "fc/config.h"
@@ -39,16 +38,17 @@ extern "C" {
     #include "flight/mixer.h"
     #include "flight/pid.h"
     #include "flight/servos.h"
-    #include "interface/cli.h"
-    #include "interface/msp.h"
-    #include "interface/msp_box.h"
-    #include "interface/settings.h"
     #include "io/beeper.h"
     #include "io/ledstrip.h"
     #include "io/osd.h"
     #include "io/serial.h"
     #include "io/vtx.h"
+    #include "msp/msp.h"
+    #include "msp/msp_box.h"
+    #include "pg/pg.h"
+    #include "pg/pg_ids.h"
     #include "pg/beeper.h"
+    #include "pg/rx.h"
     #include "rx/rx.h"
     #include "scheduler/scheduler.h"
     #include "sensors/battery.h"
