+++ src/test/unit/telemetry_crsf_msp_unittest.cc
@@ -47,10 +47,10 @@ extern "C" {
     #include "fc/config.h"
     #include "flight/imu.h"
 
-    #include "interface/msp.h"
-
     #include "io/serial.h"
     #include "io/gps.h"
+
+    #include "msp/msp.h"
 
     #include "rx/rx.h"
     #include "rx/crsf.h"
